Move APIs to separate package
